+++ pkgs/ffigen/lib/src/code_generator/imports.dart
@@ -45,8 +45,17 @@ class ImportedType extends Type {
   final String nativeType;
   final String? defaultValue;
 
-  ImportedType(this.libraryImport, this.cType, this.dartType, this.nativeType,
-      [this.defaultValue]);
+  /// Whether the [dartType] is an import from the [libraryImport].
+  final bool importedDartType;
+
+  ImportedType(
+    this.libraryImport,
+    this.cType,
+    this.dartType,
+    this.nativeType, {
+    this.defaultValue,
+    this.importedDartType = false,
+  });
 
   @override
   String getCType(Writer w) {
@@ -55,7 +64,14 @@ class ImportedType extends Type {
   }
 
   @override
-  String getFfiDartType(Writer w) => cType == dartType ? getCType(w) : dartType;
+  String getFfiDartType(Writer w) {
+    if (importedDartType) {
+      w.markImportUsed(libraryImport);
+      return '${libraryImport.prefix}.$dartType';
+    } else {
+      return cType == dartType ? getCType(w) : dartType;
+    }
+  }
 
   @override
   String getNativeType({String varName = ''}) => '$nativeType $varName';
@@ -107,30 +123,42 @@ final self = LibraryImport('self', '');
 
 final voidType = ImportedType(ffiImport, 'Void', 'void', 'void');
 
-final unsignedCharType =
-    ImportedType(ffiImport, 'UnsignedChar', 'int', 'unsigned char', '0');
+final unsignedCharType = ImportedType(
+    ffiImport, 'UnsignedChar', 'int', 'unsigned char',
+    defaultValue: '0');
 final signedCharType =
-    ImportedType(ffiImport, 'SignedChar', 'int', 'char', '0');
-final charType = ImportedType(ffiImport, 'Char', 'int', 'char', '0');
-final unsignedShortType =
-    ImportedType(ffiImport, 'UnsignedShort', 'int', 'unsigned short', '0');
-final shortType = ImportedType(ffiImport, 'Short', 'int', 'short', '0');
-final unsignedIntType =
-    ImportedType(ffiImport, 'UnsignedInt', 'int', 'unsigned', '0');
-final intType = ImportedType(ffiImport, 'Int', 'int', 'int', '0');
-final unsignedLongType =
-    ImportedType(ffiImport, 'UnsignedLong', 'int', 'unsigned long', '0');
-final longType = ImportedType(ffiImport, 'Long', 'int', 'long', '0');
+    ImportedType(ffiImport, 'SignedChar', 'int', 'char', defaultValue: '0');
+final charType =
+    ImportedType(ffiImport, 'Char', 'int', 'char', defaultValue: '0');
+final unsignedShortType = ImportedType(
+    ffiImport, 'UnsignedShort', 'int', 'unsigned short',
+    defaultValue: '0');
+final shortType =
+    ImportedType(ffiImport, 'Short', 'int', 'short', defaultValue: '0');
+final unsignedIntType = ImportedType(
+    ffiImport, 'UnsignedInt', 'int', 'unsigned',
+    defaultValue: '0');
+final intType = ImportedType(ffiImport, 'Int', 'int', 'int', defaultValue: '0');
+final unsignedLongType = ImportedType(
+    ffiImport, 'UnsignedLong', 'int', 'unsigned long',
+    defaultValue: '0');
+final longType =
+    ImportedType(ffiImport, 'Long', 'int', 'long', defaultValue: '0');
 final unsignedLongLongType = ImportedType(
-    ffiImport, 'UnsignedLongLong', 'int', 'unsigned long long', '0');
+    ffiImport, 'UnsignedLongLong', 'int', 'unsigned long long',
+    defaultValue: '0');
 final longLongType =
-    ImportedType(ffiImport, 'LongLong', 'int', 'long long', '0');
+    ImportedType(ffiImport, 'LongLong', 'int', 'long long', defaultValue: '0');
 
-final floatType = ImportedType(ffiImport, 'Float', 'double', 'float', '0.0');
-final doubleType = ImportedType(ffiImport, 'Double', 'double', 'double', '0.0');
+final floatType =
+    ImportedType(ffiImport, 'Float', 'double', 'float', defaultValue: '0.0');
+final doubleType =
+    ImportedType(ffiImport, 'Double', 'double', 'double', defaultValue: '0.0');
 
-final sizeType = ImportedType(ffiImport, 'Size', 'int', 'size_t', '0');
-final wCharType = ImportedType(ffiImport, 'WChar', 'int', 'wchar_t', '0');
+final sizeType =
+    ImportedType(ffiImport, 'Size', 'int', 'size_t', defaultValue: '0');
+final wCharType =
+    ImportedType(ffiImport, 'WChar', 'int', 'wchar_t', defaultValue: '0');
 
 final objCObjectType =
     ImportedType(objcPkgImport, 'ObjCObject', 'ObjCObject', 'void');